Move alert component to flash message dir
