+++ resources/js/components/FlashMessages/Alert.jsx
